refactor: merge keyword and variable naming checkers
Continue moving the linter towards a single AST visitor. `KeywordNamingChecker`
and `VariableNamingChecker` both walked every file and shared most of their
dispatch table (`visit_Keyword`, `visit_KeywordCall`, `visit_If`). They are now
merged into a single `NamesChecker` in the new `names.py` module, and the
decision logic moved into `check()` methods on the rules themselves.

This reduces the number of AST visitors from 25 to 24. Linter output is
unchanged.
